routes appear to be working and functioning without crashing
